% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lasUtil.R
\name{getLasFileExtent}
\alias{getLasFileExtent}
\title{Compute the bounding box of a lasFile using lasinfo}
\usage{
getLasFileExtent(lasFile)
}
\arguments{
\item{lasFile}{- The las filename}
}
\value{
The bounding box of the lasFile
}
\description{
Compute the bounding box of a lasFile using lasinfo
}
\examples{
lasFile=paste0(system.file(package='LASextractor'), '/extdata/points.las')
  lasExt = getLasFileExtent(lasFile)
  # Basic test of known dataset properties
  stopifnot(all(round(c(lasExt)) == c(680676, 3062544,  681076, 3062882)))
}

