% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getLasFileExtentPolygon}
\alias{getLasFileExtentPolygon}
\title{Compute the lasfile extent as a SpatialPolygons object}
\usage{
getLasFileExtentPolygon(lasFile, CRSinfo)
}
\arguments{
\item{lasFile}{- the las filename}

\item{CRSinfo}{- CRS object giving the coordinate system of the las data (and output polygon)}
}
\value{
-- The extent of the lasFile as a SpatialPolygons object
}
\description{
Compute the lasfile extent as a SpatialPolygons object
}
\examples{
lasFile=paste0(system.file(package='LASextractor'), '/extdata/points.las')
  # CRS of the data
  myCRS=CRS('+init=epsg:26914')
  # Get extent as a SpatialPolygons object
  mypol=getLasFileExtentPolygon(lasFile,myCRS)
  plot(mypol,axes=TRUE)
  emp = extent(mypol)
  # Basic test
  stopifnot(all(as.matrix(emp)==getLasFileExtent(lasFile)))
}

