% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LASextractor}
\alias{LASextractor}
\title{LASextractor}
\description{
Codes for getting lidar data from las files into R, and some particular routines for
interactively extracting elevation transects along lines [e.g. ditches,
riverwalls, dry channel beds,etc] \cr
 \cr
You must have have the GPL programs las2txt and lasinfo (or las2txt.exe and
lasinfo.exe on windows) installed + in your search path \cr
}
\details{
The main routines for users are (probably): \cr
 \cr
 readLas (read points from a las file); \cr
 \cr
 getLasFileExtent (get the bounding box of a .las file); \cr
 \cr
 getLasPointsNearRegion (get points from [multiple] las files in a
user-defined polygon, or within a set distance of a line);  \cr
 \cr
 lasPtsNearLine (get points from [multiple] las files near a line, and also
return information on the distance of each point along the line);  \cr
 \cr
 selectLasPts (interactively get subset of SpatialPointsDataFrame [output
from lasPtsNearLine] along a linear feature, based on a distance-elevation
curve plot); \cr
 \cr
 manuallyProjectLasElevationsAlongLine (combines the above 2 routines, can
be used to get elevation transects along linear features) \cr
 \cr
 gridXYZ (simple conversion of xyz to a raster, using an unstructured
interpolation routine) \cr
\cr
--\cr
The data in the examples was sourced from the OpenTopography Facility, with
acknowledgements below. The data in the package is a subset of the full
dataset: \cr
\cr
Dataset Name: Corpus Cristi, TX: Influence of Dunes & Barrier Islands on
Hurricane Surge(TX11_Ferreira), doi: 10.5069/G9862DC4, Survey Date 10/17/2011
\cr
 \cr
Dataset Acknowledgement: LiDAR data acquisition and processing completed by
the National Center for Airborne Laser Mapping (NCALM - http://www.ncalm.org).
NCALM funding provided by NSF's Division of Earth Sciences, Instrumentation and
Facilities Program. EAR-1043051.\cr
 \cr
Data Access Acknowledgement: This material is based on [data, processing]
services provided by the OpenTopography Facility with support from the National
Science Foundation under NSF Award Numbers 0930731 & 0930643 \cr
}
\author{
Gareth Davies \email{gareth.davies.ga.code@gmail.com}
}

