% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readLas}
\alias{readLas}
\title{Convert points in las file to matrix using las2txt.}
\usage{
readLas(lasFile, var = "xyzc", verbose = FALSE, xRange = NULL,
  yRange = NULL, zRange = NULL)
}
\arguments{
\item{lasFile}{-- las filename}

\item{var}{-- flags for which variables we return, see las2txt help for
info. The first 3 letters must be xyz}

\item{verbose}{-- TRUE/FALSE print lots of information}

\item{xRange}{-- length 2 numeric vector: Only extract points within these x values}

\item{yRange}{-- length 2 numeric vector: Only extract points within these y values}

\item{zRange}{-- length 2 numeric vector: Only extract points within these z values}
}
\value{
Matrix with all var variables. These variables must be numeric
}
\description{
The function uses a simple system call to las2txt, so requires that software
to be installed + in your search path
}
\examples{
lasFile=paste0(system.file(package='LASextractor'), '/extdata/points.las')
  las_xyz=readLas(lasFile)
  summary(las_xyz)
  # Check the data size + basic properties
  stopifnot(all(dim(las_xyz)==c(744826,4)))
  stopifnot(round(mean(las_xyz[,1]))==680847)
  stopifnot(round(mean(las_xyz[,2]))==3062726)
}

