% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lasUtil.R
\name{selectLasPts}
\alias{selectLasPts}
\title{Interactively select a subset of las points}
\usage{
selectLasPts(shpFile, verticalThreshold, subplotLength = Inf)
}
\arguments{
\item{shpFile}{-- [as output from lasPtsNearLine] a point shapefile
filename, OR a SpatialPointsDataFrame
object with the same information. Attribute table has 'elevation' in first
column, and 'distance' in the last column}

\item{verticalThreshold}{-- returned points will be within
'verticalThreshold' elevation distance of the interactively defined
distance-elevation function.}

\item{subplotLength}{-- Optionally the digitization occurs on a sequence
of distance-elevation plots, where each plot covers a horizontal length of
approximately subplotLength (physical distance in metres). This is needed to
digitize long features}
}
\value{
A list containing 1): The SpatialPointsDataFrame containing points in
   shpFile that are within verticalThreshold of the clicked points;
   2) The distance-elevation curve that the user clicked off
}
\description{
Interactively extract subset of spatialPointsDataFrame (output from
lasPtsNearLine) along a feature, using its distance-elevation plot
}
\details{
Suppose the input xyz points roughly follow a feature of interest called a
'riverwall', which is distinguished from the surrounding topography by its
elevation [realistic examples: a riverwall, a narrow drainage channel, a
ditch, an elevated road, a bridge, or similar].
Some input xyz points are on the riverwall
, while others are beside it, and the difference is obvious from
the point elevation. Distances of input xyz points along the riverwall are
assumed known (see lasPtsNearLine). \cr
We want to extract those input points which are along the riverwall. \cr
To achieve this the input points are plotted as along-wall-distance-vs-elevation, and
the user interactively adds a distance-elevation-curve to the plot.
Finally, we return all input xyz points with |elevation -
distance-elevation-curve| < verticalThreshold, and the distance elevation curve.
}
\examples{
\dontrun{
    lasFile=paste0(system.file(package='LASextractor'), '/extdata/points.las')
    lineShpFile=paste0(system.file(package='LASextractor'),'/extdata/Linear_extraction_line.shp')
    lasSubset=lasPtsNearLine(lineShpFile, lasFile, bufWidth=10., outdir=tempdir())
    # Click off points on a plot, and return all points within 20cm of line defined by clicks
    newPts = selectLasPts(lasSubset,verticalThreshold=0.2)
}
}

